backend/requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
pydantic==2.12.5
python-dateutil==2.9.0.post0
python-multipart==0.0.24
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
google-cloud-storage==2.18.2
